% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpR2-package.R
\docType{package}
\name{elpR2-package}
\alias{elpR2}
\alias{elpR2-package}
\title{elpR2: The Elephant Listening Project File Management}
\description{
Tools for processing and analyzing elephant rumble detector output, gunshot detector output, and other detector output in Raven Pro selection tables. Functions include selection table restructuring, sound quality checking, detector performance evaluation, and detection summaries for acoustic monitoring data.
}
\author{
\strong{Maintainer}: Bobbi J. Estabrook \email{bobbi.estabrook@cornell.edu}

Other contributors:
\itemize{
  \item Jidapa Janpathompong [contributor]
}

}
\keyword{internal}
