% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_function.R
\name{maps_function}
\alias{maps_function}
\title{Maps}
\usage{
maps_function(folder_path, site_lat_long)
}
\arguments{
\item{folder_path}{Path to folder containing output from data_summaries_function(). ***Needs to be in /data_summaries/summary_tables if running this function from shiny app in elpRApp().}

\item{site_lat_long}{Path to a .txt file containing site names with latitude, longitude, strata (optional), vegetation class (optional): See existing files in the "elpR/Files/sites" folder for required structure of .txt file}
}
\value{
This function outputs a list containing tables and leaflet maps:
\itemize{
\item event_site_avg: a table summarizing average daily events per site
\item event_year_avg: a table summarizing average daily events per site per year
\item event_month_avg: a table summarizing average daily events per site per month
\item default_map: a leaflet map displaying average daily events at each site
\item strata_map: a leaflet map displaying average daily events at each site, with sites colored by strata
\item veg_map: a leaflet map displaying average daily events at each site, with sites colored by vegetation class
\item year_map: a leaflet map with sites colored by average daily events; data intended to be displayed in elpRApp()
\item month_map: a leaflet map with sites colored by average daily events; data intended to be displayed in elpRApp()
}
}
\description{
This function reads tables generated by data_summaries_function() and
creates maps to be visualized in elpRApp().
}
\author{
Jidapa Janpathompong
}
