% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_gunshot_function.R
\name{restructure_gunshot_function}
\alias{restructure_gunshot_function}
\title{Gunshot Selection Table Restructure}
\usage{
restructure_gunshot_function(
  parent_dir,
  deployment_name,
  deployment_num,
  disk_ID,
  sites,
  sample_rate,
  Detector,
  Detector_ScoreThreshold,
  Filter_ScoreThreshold
)
}
\arguments{
\item{parent_dir}{Folder containing the specific elpR sub-folder structure}

\item{deployment_name}{Deployment name (e.g., nn_202302_feb)}

\item{deployment_num}{Deployment number (numeric)}

\item{disk_ID}{Disk ID (name of the disk with your sound files)}

\item{sites}{A .txt file containing site names with a header of "Site" (See existing files in the "elpR/Files/sites" folder for required structure of .txt file)HH_selection_tables Do not change this value}

\item{sample_rate}{Sample rate of the sound files that the detector was run on}

\item{Detector}{For the gunshot data template detector:"DTDguns8"}

\item{Detector_ScoreThreshold}{Score that the detector was run with (e.g., 0.53 for DTDguns8)}

\item{Filter_ScoreThreshold}{Score that the final processed tables should be filtered by (e.g., 0.53 for DTDguns8)}
}
\value{
Output files located in:
\itemize{
\item /Selection_Tables/gunshot/final
\item /num_events/gunshot
\item /zero_days_SSTs/gunshot
\item A new tab in the associated /sound_check .xlsx file.
}
}
\description{
This function will:
\itemize{
\item 1) Run RavenX selection tables through the R script (files can be in subfolders)
\item 2) Concatenate the tables
\item 3) Filter the saved selection tables by score and length using (keep copies of both the filtered and unfiltered files)
}
}
\note{
This function requires:
\itemize{
\item Detector output from the gunshot data template detector
\item All detector output saved to one folder (no site sub-folders)
\item Selection tables that are a max length of 10000 selections, with multiple sites merged into one, and minimum score threshold of 0.53
\item num_events file created by the DTD app
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
