% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_sounds_function.R
\name{sound_exclude_function}
\alias{sound_exclude_function}
\title{Exclude Bad Sounds}
\usage{
sound_exclude_function(
  sound_path,
  parent_dir,
  extra_sounds_folder,
  deployment_name,
  deployment_num,
  disk_ID,
  sound_file_ext,
  have_SwiftFiles,
  merge_swift_files
)
}
\arguments{
\item{sound_path}{Folder containing sound files}

\item{parent_dir}{Folder containing the specific elpR sub-folder structure}

\item{extra_sounds_folder}{Folder to output bad sound files}

\item{deployment_name}{Deployment name (e.g., nn_202302_feb)}

\item{deployment_num}{Deployment number (numeric)}

\item{disk_ID}{Disk ID (name of the disk with your sound files)}

\item{sound_file_ext}{Sound file extension type (e.g., ".wav", ".aif")}

\item{have_SwiftFiles}{If the user has swift files in the sound folders (y/n)}

\item{merge_swift_files}{If the user wants the swift files to be merged into one file (y/n). For this to work, all the swift files must have the same structure (columns names and order)}
}
\value{
Output files
\itemize{
\item Summary of the duration of the excluded files and save the info as a new tab in the sound_check xlsx file
\item Move the Swift configuration and debug .txt files to the swift_files folder (/swift_files) if the user specifies that those files exist in the sound folders
\item Merge the Swift Config files into one .txt file and append that to the sound check sheet (if the user specifies that they want the merge and have the Swift files)
}
}
\description{
This function will move sound files that are marked as 'bad' sounds in the sound check spreadsheet to another folder.
The function will:
\itemize{
\item Read the "sounds" worksheet in the sound check excel file that is produced by the sound_check script and function. The sound_check file must exist for this function to work and be in the following directory in the elpR File folder: /sound_check
\item Move all sounds marked with a 'y' in the "Exclude (y)" column of the sound_check.xlsx file to the excluded sounds folder defined in the script.
\item Sound files to exclude include files with less than 1 min of recording and bad sound files (automatically marked as 'y" by the sound.check function). If the user manually marked bad sound files or sounds to exclude with a 'y' in the Exclude (y/n) column, it will remove those sound files.
\item For elephant rumbles, excluded sounds are those with less than 22.95 hours or more than 24 hours (automatically marked by the sound.check function).
\item Summarize the duration of the excluded files and save the info as a new tab in the sound_check.xlsx file.
\item Merge the Swift Config files into one .txt file and append that to the sound check sheet (if the user specifies that they want the merge and have the Swift files).
\item This will add the site name to Swift .txt files and move them out of the site folder to the ~/swift_files folder if have_SwiftFiles was set to “y”.
}
}
\note{
Run this step AFTER running the sound.check function and BEFORE running the detectors on the sound files
The user should confirm in the sound_check excel file which sounds are marked as 'y' in the Exclude (y/n) column to be sure those sounds should be excluded and to add 'y' to sounds that should be excluded but were not marked with a 'y' automatically by the sound.check function.
DO NOT rename the sound_check Excel file after you run the sound.check function. The scripts in this package often look for the sound_check file by the deployment name, deployment number, and disk number that the user specifies.
The 'sound check' file is required for this function. The name of the file should be unchanged from when it was created. Use the same variable names between the two scripts for each project, deployment, and disk
\itemize{
\item WARNING: New files will overwrite old files with the same name. Be sure to move files of the same name out of the sound folder before running this script if you don't want to overwrite them
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
