% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summaries_function.R
\name{data_summaries_function}
\alias{data_summaries_function}
\title{Data Summaries}
\usage{
data_summaries_function(
  parent_dir,
  project_name,
  deployment_num,
  detector_name,
  ele_tables,
  zero_txt,
  sound_check_include,
  sound_checks,
  use_only_sites_provided,
  site_lat_long,
  rand_dates_needed,
  ele_bad_sound_remove
)
}
\arguments{
\item{parent_dir}{Folder containing the specific elpR sub-folder structure}

\item{project_name}{Project name (e.g., PNNN, DzangaBai, Kakum)}

\item{deployment_num}{Deployment(s) number (e.g., 01-12, 04)}

\item{detector_name}{Detector used (e.g., HoriHarm, FruitPunchAI, Stanford Detector)}

\item{ele_tables}{Folder containing selection tables}

\item{zero_txt}{Folder containing zero-days files generated by restrucutre_rumble_function()}

\item{sound_check_include}{If you have sound check .xlsx files for this script to filter out bad sounds (y/n)}

\item{sound_checks}{Folder containing sound check .xlsx files}

\item{use_only_sites_provided}{If you want to include only the sites listed in a site_lat_long file (y/n)}

\item{site_lat_long}{A .txt file containing site names with latitude, longitude, vegetation class (optional), strata (optional): See existing files in the "elpR/Files/sites" folder for required structure of .txt file}

\item{rand_dates_needed}{If you want 3 random days per week (y/n) *dates are pre-selected in the /rand-days/Rand_dates.csv file}

\item{ele_bad_sound_remove}{If you want to exclude sounds with less than 23 hrs of sound (y/n)}
}
\value{
This function outputs:
\itemize{
\item Summary tables
\itemize{
\item Hourly summary tables include:
\itemize{
\item Site
\item Date
\item File Name
\item Hour
\item Deployment Number
\item Sum Rumbles
\item Latitude
\item Longitude
\item Vegetation Class (if provided in sites file)
\item Strata (if provided in sites file)
\item Month
\item Year
\item WeekDate
\item DayNight: If the hour is considered day (06:00 – 18:00) or night (18:00 – 06:00)
}
\item Daily summary tables include:
\itemize{
\item Date
\item Site
\item WeekDate
\item Latitude
\item Longitude
\item Vegetation Class (if provided in sites file)
\item Strata (if provided in sites file)
\item total_rumbles
\item n_hours: Total hours sampled for the date
\item n_valid_hours: Total valid hours sampled for the date
\item night_total: Total nighttime rumbles for the date
\item prop_night: Proportion of nighttime rumbles to all rumbles for the date
\item prop_day: Proportion of daytime rumbles to all rumbles for the date
\item Month
\item Year
\item Week
}
\item Weekly summary tables include:
\itemize{
\item WeekDate
\item Site/Strata/Vegetation Class
\item Latitude
\item Longitude
\item Vegetation Class (if provided in sites file)
\item Strata (if provided in sites file)
\item MeanDailyRumbles: Mean number of rumbles per day for the week
\item sdDailyRumbles: Standard deviation of the number of rumbles per day for the week
\item n: Number of sampled days included in weekly calculations
\item seDailyRumbles: Standard error of the number of rumbles per day for the week
\item maxDailyRumbles: Maximum number of rumbles per day for the week
\item minDailyRumbles: Minimum number of rumbles per day for the week
\item sumRumbles: Total number of rumbles for the week
\item sumRumblesNight: Total number of nighttime rumbles for the week
\item meanRumblesNight: Mean number of nighttime rumbles for the week
\item sdRumblesNight: Standard deviation of the number of nighttime rumbles for the week
\item seRumblesNight: Standard error of the number of nighttime rumbles for the week
\item PropNightRumbles: Proportion of nighttime rumbles to all rumbles for the week
\item SumRumblesPerDaysSampled: Total daytime rumbles for the week
\item SumNightRumblesPerDaySampled: Nighttime rumbles divided by number of days per week
}
\item Monthly summary tables include:
\itemize{
\item Year
\item Month
\item MeanDailyRumbles: Mean number of rumbles per day for the month
\item sdDailyRumbles: Standard deviation of the number of rumbles per day for the month
\item n: Number of sampled days included in monthly calculations
\item seDailyRumbles: Standard error of the number of rumbles per day for the month
\item maxDailyRumbles: Maximum number of rumbles per day for the month
\item minDailyRumbles: Minimum number of rumbles per day for the month
\item sumRumbles: Total number of rumbles for the month
\item sumNightRumbles: Total number of nighttime rumbles for the month
\item meanNightRumbles: Mean number of nighttime rumbles for the month
\item proportionNightRumbles: Proportion of nighttime rumbles to all rumbles for the month
\item Year-Month: WeekDate
}
}
\item Summary plots
\itemize{
\item Plots saved in "saved_plots.Rds" include:
\itemize{
\item total_plot: Line plot of total rumbles per hour per site
\item night_day_plot: Box plot of total daytime and nighttime rumbles for all sites
\item hourly_plot: Bar plot of mean rumbles per hour for all sites
\item daily_plot_site: Line plots of total rumbles per date by site
\item daily_plot_strata: Line plots of total rumbles per date by strata (if strata is provided in sites file)
\item daily_plot_veg: Line plots of total rumbles per date by vegetation class (if vegetation class is provided in sites file)
\item weekly_plot_site: Line plots of mean number of daily rumbles per week per site
\item weekly_plot_strata: Line plots of mean number of daily rumbles per week per strata (if strata provided in sites file)
\item weekly_plot_veg: Line plots of mean number of daily rumbles per week per vegetation class (if vegetation class is provided in sites file)
\item monthly_plot: Line plot of mean number of daily rumbles per month for all sites
\item monthly_plot_site: Line plots of mean number of daily rumbles per month by site
\item monthly_plot_strata: Line plots of mean number of daily rumbles per month by strata (if strata is provided in sites file)
\item monthly_plot_strata_overlay: Line plots from monthly_plot_strata on the same graph
}
}
}
}
\description{
This function summarizes data from the annotated selection tables, the sound check files, and the zero-days selection tables into hourly, daily, weekly, and monthly time periods.
The summaries will yield .txt files containing summary data and an .rds file containing plots.
This function will:
\itemize{
\item Install and load necessary packages
\item Read in data provided by the user (sound check files, random dates, selection tables, zero-day tables, sites file)
\item Generate summary tables and plots that are saved to /data_summaries/summary_tables and /data_summaries/summary_plots, respectively
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
