% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpR_shiny_app.R
\name{elpRApp}
\alias{elpRApp}
\title{Shiny App for elpR}
\usage{
elpRApp()
}
\value{
A shiny app
}
\description{
This function creates a shiny app that allows users to run the other
functions in this package (sound_check_function(), exclude_sounds_function(),
restructure_rumble_function(), restructure_gunshot_function(),
restructure_general_function(), and data_summaries_function()), and visualize
their data.
}
\author{
Jidapa Janpathompong
}
