% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_rumble_function.R
\name{restructure_rumble_function}
\alias{restructure_rumble_function}
\title{Rumble Selection Table Restructure}
\usage{
restructure_rumble_function(
  parent_dir,
  deployment_name,
  deployment_num,
  disk_ID,
  sites,
  sample_rate,
  three_rand_days,
  min_23hrs,
  score_column_name,
  Detector,
  Detector_ScoreThreshold,
  Filter_ScoreThreshold
)
}
\arguments{
\item{parent_dir}{Folder containing the specific elpR sub-folder structure}

\item{deployment_name}{Deployment name (e.g., nn_202302_feb)}

\item{deployment_num}{Deployment number (numeric)}

\item{disk_ID}{Disk ID (name of the disk with your sound files)}

\item{sites}{A .txt file containing site names with a header of "Site" (See existing files in the "elpR/Files/sites" folder for required structure of .txt file)HH_selection_tables Do not change this value}

\item{sample_rate}{Sample rate of the sound files that the detector was run on}

\item{three_rand_days}{If you want 3 random days per week (y/n) *dates are pre-selected in the /rand-days/Rand_dates.csv file}

\item{min_23hrs}{If you want to exclude sounds with less than 23 hrs of sound (y/n)}

\item{score_column_name}{Score column name (e.g., "Confidence" in BirdNET, "Score" in Hori-Harm)}

\item{Detector}{For HoriHarm: HHv6, FruitPunchAI: FPv1, Stanford Detector: SDv1}

\item{Detector_ScoreThreshold}{Score that the detector was run with (e.g., 0.2 for Hori-Harm rumbles)}

\item{Filter_ScoreThreshold}{Score that the final processed tables should be filtered by (e.g., 0.4 for Hori-Harm rumbles)}
}
\value{
Output files
\itemize{
\item Updated sound check file with new tab called Rumble Detector Summaries (/sound_check)
\itemize{
\item Duration of good sounds per site
\item Total detections at lowest score per site
\item Total detections above score threshold per site
\item Total detections above score threshold and on random dates per site (will be blank if random dates are not enabled by user)
}
\item Raven Pro sound selection tables (/Selection_Tables/rumble/final) in folders by score
\item Empty selection tables. These are dummy selections tables that have 1 selection per sound file which did not have a detection (/Empty_Tables/rumble) greater than the initial detector score
\item Summary of the number of events per sound file (/num_events/rumble)
\item Plots for the number of detections per day per site by the detector score and the filtered score (/num_events/rumble)
\item Dummy selection tables with selections on sound files that did not have a detection above the filtered score (/zero_days_SSTs/rumble)
}
}
\description{
This function will:
\itemize{
\item Read in all selection tables from elephant rumble detector output in the /Selection_Tables/rumble/raw folder (required that they are in site-folders)
\item Restructure those tables by merging, cleaning, and adding new columns
\item Merge the detector tables into site-wise selection tables
\item Add 'sound check' info to the selection tables (the 'sound_check' function should have been run on associated sound files BEFORE this step)
\item Add the random days designation to the 3 randomly selected days per week (user defined by parameter three_rand_days)
\item Ignore detections on sound files that were marked with 'y' (bad sound file) in the 'Exclude (y/n)' column of the 'Sound Check' Excel file
\item If min_23hrs <- "y" then ignore detections on sound files that were marked with 'e' (bad sound file for elephant analysis) in the 'Exclude (y/n)' column of the 'Sound Check' Excel file
\item If three_rand_days <- "y" then summarize and produce selections tables for the 3 random days per week
\item Save selection tables that are filtered by score per table (in the /Selection_Tables/rumble/final folder)
\item Count the total number of detections per site and day by score and sound file duration (added to the 'sound check' Excel sheet as a tab, saved in the /sound_check folder)
\item Calculates the total number of detections per sound file saved as a .txt file in the /num_events/rumble? folder
\item Plot the total detections per score as .png files (saved in /num_events/rumble folder)
\item Count the total number of days without detections
\item Save a selection table of sound files without detections at threshold score (user defined) for random dates (saved in the /zero_days_SSTs/rumble folder)
\item Save a .txt file with list of selection tables that had 0 detections for record keeping
\item If the detector specified is "FPv1", the script will reset the minimum frequency of the detections to 10 Hz, because the FruitPunch detector (FPv1) generates minimum frequencies that are higher than the detected signal.
}
}
\note{
\itemize{
\item Input detector files require Raven Pro columns:
\itemize{
\item Begin File
\item File Offset (s)
\item Begin Time (s)
\item End Time (s)
\item High Freq (Hz)
\item Low Freq (Hz)
\item Score (can be named something else, but must be indicated by user in script)}
\item The 'sound check' file is required for this function. The name of the file should be unchanged from when it was created. Use the same variable names between the two scripts for each project, deployment, and disk.
This script will only generate selection tables and summaries for sites that are included in the corresponding 'sound check' file.
\item WARNING: New files will overwrite old files with the same name. Be sure to move files of the same name out of the 'Packages/elpR/File/Rumble/Final' folder before running this script if you don't want to overwrite them
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
